requirements.txt
python-telegram-bot~=22.0
python-dotenv~=1.0.1
youtube-transcript-api~=1.0.3
litellm~=1.67.0
SQLAlchemy~=2.0.40
pyfreeproxies~=0.2.2